+++ tests/integration/bulkexports/v1/export/test_export_custom_job.py
@@ -74,7 +74,9 @@ class ExportCustomJobTestCase(IntegrationTestCase):
                         "end_day": "end_day",
                         "webhook_url": "webhook_url",
                         "email": "email",
-                        "resource_type": "resource_type"
+                        "resource_type": "resource_type",
+                        "job_queue_position": "1",
+                        "estimated_completion_time": "2021-03-15T20:20:14.547"
                     }
                 ]
             }
@@ -114,7 +116,9 @@ class ExportCustomJobTestCase(IntegrationTestCase):
                 "end_day": "end_day",
                 "webhook_url": "webhook_url",
                 "email": "email",
-                "resource_type": "resource_type"
+                "resource_type": "resource_type",
+                "job_queue_position": "1",
+                "estimated_completion_time": "2021-03-15T20:20:14.547"
             }
             '''
         ))