+++ tests/integration/video/v1/test_room.py
@@ -150,6 +150,44 @@ class RoomTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
+    def test_create_group_rooms_response(self):
+        self.holodeck.mock(Response(
+            201,
+            '''
+            {
+                "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "date_created": "2015-07-30T20:00:00Z",
+                "date_updated": "2015-07-30T20:00:00Z",
+                "status": "in-progress",
+                "type": "group",
+                "sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "enable_turn": true,
+                "unique_name": "grouproom",
+                "max_participants": 50,
+                "max_concurrent_published_tracks": 10,
+                "duration": 0,
+                "status_callback_method": "POST",
+                "status_callback": "",
+                "record_participants_on_connect": false,
+                "video_codecs": [
+                    "VP8"
+                ],
+                "media_region": "us1",
+                "end_time": "2015-07-30T20:00:00Z",
+                "url": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "participants": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
+                    "recordings": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Recordings",
+                    "recording_rules": "https://video.twilio.com/v1/Rooms/RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/RecordingRules"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.video.v1.rooms.create()
+
+        self.assertIsNotNone(actual)
+
     def test_list_request(self):
         self.holodeck.mock(Response(500, ''))